+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 Initial release of chef-security-essentials. 
 
 Features:
+
 * fail2ban
 * sysadmin users and group
 * openssh restrictions
 * basic firewall
+* unattended upgrades
+* log notifications